[GraphMaker] Refactor script
