+++ GraphMaker/README.md
@@ -8,3 +8,4 @@ When writing python code, please follow the given rules:
 - variable and method names are in `lower_case` (refer to PEP8)
 - constants are `ALL_UPPER_CASE`
 - `mixedCase` is **not** welcome in Python code
+
