+++ GraphMaker/app/data/NodeData.py
@@ -1,0 +1,4 @@
+class NodeData:
+    def __init__(self, name, aliases):
+        self.name, self.aliases = name, aliases
+
