+++ GraphMaker/app/general/constants.py
@@ -1,0 +1,18 @@
+# To be removed since everything should be handled by ElementTree
+SET_TABS_IN_XML = True
+TAB = {True: '\t', False: ''}[SET_TABS_IN_XML]
+
+EXTERNAL_EDGES_SEPARATOR = ' --- '
+
+# tkinter events defines
+LEFT_CLICK = '<Button-1>'
+WHEEL_CLICK = '<Button-2>'
+RIGHT_CLICK = '<Button-3>'
+WHEEL_UP = '<Button-4>'
+WHEEL_DOWN = '<Button-5>'
+LEFT_RELEASE = '<ButtonRelease-1>'
+WHEEL_RELEASE = '<ButtonRelease-2>'
+RIGHT_RELEASE = '<ButtonRelease-3>'
+LEFT_CLICK_MOTION = '<B1-Motion>'
+WHEEL_CLICK_MOTION = '<B2-Motion>'
+RIGHT_CLICK_MOTION = '<B3-Motion>'
