+++ GraphMaker/app/general/tkinter_imports.py
@@ -1,0 +1,4 @@
+import tkinter as t
+from tkinter import filedialog as fdialog
+from tkinter.simpledialog import askfloat
+from tkinter import messagebox as mbox
